Optimize the cache updater to minimize disk accesses
The bottleneck is disk accesses, so we rewrote the updater to to
minimize them. We trade higher memory for reduced disk accesses. We:

1. Execute one big SQL query at the start to fetch the relevant previous caches.
2. Skip reading a file's data if the mtime has not changed since the previous cache update.
3. Only execute a SQLite upsert if the read data differ from the previous caches.

With these optimizations, we make a lot of readdir and stat calls, but
minimize file and database accesses to solely the files that have
updated since the last cache run.